refactor: 拆分 Utils 檔案下的函式
